requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
django-import-export==3.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
